requirements.txt
flask
discord.py
python-dotenv
pytesseract
Pillow
